Add enum members
