+++ src/EditorFeatures/CSharpTest/Completion/CompletionProviders/CompletionProviderOrderTests.cs
@@ -47,7 +47,6 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.Completion.CompletionPr
                 typeof(ExplicitInterfaceTypeCompletionProvider),
                 typeof(ObjectCreationCompletionProvider),
                 typeof(ObjectAndWithInitializerCompletionProvider),
-                typeof(EnumCompletionProvider),
                 typeof(CSharpSuggestionModeCompletionProvider),
                 typeof(EnumAndCompletionListTagCompletionProvider),
                 typeof(CrefCompletionProvider),
