+++ src/Features/CSharp/Portable/Completion/CompletionProviders/EnumAndCompletionListTagCompletionProvider.cs
@@ -2,9 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System;
+using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Composition;
 using System.Linq;
@@ -20,6 +19,7 @@ using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.LanguageServices;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Shared.Extensions;
+using Microsoft.CodeAnalysis.Shared.Utilities;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
@@ -30,6 +30,14 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
     [Shared]
     internal partial class EnumAndCompletionListTagCompletionProvider : LSPCompletionProvider
     {
+        private static readonly CompletionItemRules s_enumTypeRules =
+            CompletionItemRules.Default.WithCommitCharacterRules(ImmutableArray.Create(CharacterSetModificationRule.Create(CharacterSetModificationKind.Replace, '.')))
+                                       .WithMatchPriority(MatchPriority.Preselect)
+                                       .WithSelectionBehavior(CompletionItemSelectionBehavior.HardSelection);
+
+        private static readonly CompletionItemRules s_enumMemberRules =
+            CompletionItemRules.Default.WithCommitCharacterRules(ImmutableArray.Create(CharacterSetModificationRule.Create(CharacterSetModificationKind.Replace, '.')));
+
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
         public EnumAndCompletionListTagCompletionProvider()
@@ -61,10 +69,9 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
             {
                 var document = context.Document;
                 var position = context.Position;
-                var options = context.Options;
                 var cancellationToken = context.CancellationToken;
 
-                var tree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
+                var tree = await document.GetRequiredSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
                 if (tree.IsInNonUserCode(position, cancellationToken))
                     return;
 
@@ -104,10 +111,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
             // If we have a Nullable<T>, unwrap it.
             if (type.OriginalDefinition.SpecialType == SpecialType.System_Nullable_T)
             {
-                type = type.GetTypeArguments().FirstOrDefault();
-
-                if (type == null)
+                var typeArg = type.GetTypeArguments().FirstOrDefault();
+                if (typeArg == null)
                     return;
+
+                type = typeArg;
             }
 
             var document = context.Document;
@@ -115,15 +123,18 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
             var semanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
             if (type.TypeKind != TypeKind.Enum)
             {
-                type = TryGetEnumTypeInEnumInitializer(semanticModel, token, type, cancellationToken) ??
-                       TryGetCompletionListType(type, semanticModel.GetEnclosingNamedType(position, cancellationToken), semanticModel.Compilation);
+                var enumType = TryGetEnumTypeInEnumInitializer(semanticModel, token, type, cancellationToken) ??
+                               TryGetCompletionListType(type, semanticModel.GetEnclosingNamedType(position, cancellationToken), semanticModel.Compilation);
 
-                if (type == null)
+                if (enumType == null)
                     return;
+
+                type = enumType;
             }
 
             var options = context.Options;
-            if (!type.IsEditorBrowsable(options.GetOption(CompletionOptions.HideAdvancedMembers, semanticModel.Language), semanticModel.Compilation))
+            var hideAdvancedMembers = options.GetOption(CompletionOptions.HideAdvancedMembers, semanticModel.Language);
+            if (!type.IsEditorBrowsable(hideAdvancedMembers, semanticModel.Compilation))
                 return;
 
             // Does type have any aliases?
@@ -133,17 +144,47 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
                 ? alias.Name
                 : type.ToMinimalDisplayString(semanticModel, position);
 
-            var item = SymbolCompletionItem.CreateWithSymbolId(
-                displayText: displayText,
-                displayTextSuffix: "",
-                symbols: ImmutableArray.Create(alias ?? type),
-                rules: s_rules.WithMatchPriority(MatchPriority.Preselect),
-                contextPosition: position);
+            // Add the enum itself.
+            var symbol = alias ?? type;
+            var sortText = symbol.Name;
 
-            context.AddItem(item);
+            context.AddItem(SymbolCompletionItem.CreateWithSymbolId(
+                displayText,
+                displayTextSuffix: "",
+                symbols: ImmutableArray.Create(symbol),
+                rules: s_enumTypeRules,
+                contextPosition: position,
+                sortText: sortText));
+
+            // And now all the accessible members of the enum.
+            if (type.TypeKind == TypeKind.Enum)
+            {
+                // We'll want to build a list of the actual enum members and all accessible instances of that enum, too
+                var index = 0;
+                foreach (var field in GetSortedEnumMembers(type))
+                {
+                    index++;
+                    if (!field.IsEditorBrowsable(hideAdvancedMembers, semanticModel.Compilation))
+                        continue;
+
+                    context.AddItem(SymbolCompletionItem.CreateWithSymbolId(
+                        displayText: $"{displayText}.{field.Name}",
+                        displayTextSuffix: "",
+                        symbols: ImmutableArray.Create<ISymbol>(field),
+                        rules: s_enumMemberRules,
+                        contextPosition: position,
+                        sortText = $"{sortText}_{index:0000}"));
+                }
+            }
         }
 
-        private static ITypeSymbol TryGetEnumTypeInEnumInitializer(
+        private static IEnumerable<IFieldSymbol> GetSortedEnumMembers(ITypeSymbol type)
+        {
+            var fields = type.GetMembers().OfType<IFieldSymbol>().Where(f => f.IsConst).Where(f => f.HasConstantValue);
+            return fields.OrderBy(f => IntegerUtilities.ToInt64(f.ConstantValue));
+        }
+
+        private static ITypeSymbol? TryGetEnumTypeInEnumInitializer(
             SemanticModel semanticModel, SyntaxToken token,
             ITypeSymbol type, CancellationToken cancellationToken)
         {
@@ -169,14 +210,17 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
                     // If so, walk back to the token before the operator token and see if it binds to a member
                     // of this enum.
                     var previousToken = token.GetPreviousToken();
-                    var symbol = semanticModel.GetSymbolInfo(previousToken.Parent, cancellationToken).Symbol;
-
-                    if (symbol?.Kind == SymbolKind.Field &&
-                        containingType.Equals(symbol.ContainingType))
+                    if (previousToken.Parent != null)
                     {
-                        // If so, then offer this as a place for enum completion for the enum we're currently 
-                        // inside of.
-                        return containingType;
+                        var symbol = semanticModel.GetSymbolInfo(previousToken.Parent, cancellationToken).Symbol;
+
+                        if (symbol?.Kind == SymbolKind.Field &&
+                            containingType.Equals(symbol.ContainingType))
+                        {
+                            // If so, then offer this as a place for enum completion for the enum we're currently 
+                            // inside of.
+                            return containingType;
+                        }
                     }
                 }
             }
@@ -187,26 +231,20 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
         protected override Task<CompletionDescription> GetDescriptionWorkerAsync(Document document, CompletionItem item, CancellationToken cancellationToken)
             => SymbolCompletionItem.GetDescriptionAsync(item, document, cancellationToken);
 
-        private static readonly CompletionItemRules s_rules =
-            CompletionItemRules.Default.WithCommitCharacterRules(ImmutableArray.Create(CharacterSetModificationRule.Create(CharacterSetModificationKind.Replace, '.')))
-                                       .WithMatchPriority(MatchPriority.Preselect)
-                                       .WithSelectionBehavior(CompletionItemSelectionBehavior.HardSelection);
-
-        private static INamedTypeSymbol TryGetCompletionListType(ITypeSymbol type, INamedTypeSymbol within, Compilation compilation)
+        private static INamedTypeSymbol? TryGetCompletionListType(ITypeSymbol type, INamedTypeSymbol? within, Compilation compilation)
         {
+            if (within == null)
+                return null;
+
             // PERF: None of the SpecialTypes include <completionlist> tags,
             // so we don't even need to load the documentation.
             if (type.IsSpecialType())
-            {
                 return null;
-            }
 
             // PERF: Avoid parsing XML unless the text contains the word "completionlist".
             var xmlText = type.GetDocumentationCommentXml();
             if (xmlText == null || !xmlText.Contains(DocumentationCommentXmlNames.CompletionListElementName))
-            {
                 return null;
-            }
 
             var documentation = CodeAnalysis.Shared.Utilities.DocumentationComment.FromXmlFragment(xmlText);
