+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Utilities/IntegerUtilities.cs
@@ -76,7 +76,7 @@ namespace Microsoft.CodeAnalysis.Shared.Utilities
             => o is ulong ? (ulong)o : unchecked((ulong)System.Convert.ToInt64(o));
 
         public static long ToInt64(object? o)
-            => o is ulong ? unchecked((long)(ulong)o) : System.Convert.ToInt64(o);
+            => o is ulong ul ? unchecked((long)ul) : System.Convert.ToInt64(o);
 
         public static bool IsIntegral([NotNullWhen(true)] object? value)
             => value switch
